+++ cli/cmd/setup.go
@@ -1,7 +1,7 @@
 package cmd
 
 import (
-	"github.com/leapfrogtechnology/shift/cli/internal/setup"
+	"github.com/leapfrogtechnology/shift/cli/internals/setup"
 )
 
 // Setup prompts user for required information for creating a project.
